refactor(image): resolve #400 code-quality cleanups
- Split cairo.rs (~900 LOC, three responsibilities) into a directory module:
  cairo/{mod,surface,context,pattern}.rs. Pure code move, logic unchanged.
- sharpenImage now reuses filter::convolve3x3 instead of hand-rolling the same
  3x3 clamp-edge convolution.
- Remove the dead duplicate branch in cairo append_arc (the `else if i == 0` and
  `else` arms were identical line_to calls).
- Recover poisoned table mutexes: lock_recover() replaces all
  `.lock().unwrap()` so that after ffi_guard catches a panic, the bridge keeps
  serving later calls instead of degrading every operation to its sentinel.

Closes #400.
